รวบรวมตัวเลือกตามหน้าต่างๆ เพื่อนำมาสร้างออร์เดอร์

+++ lib/home/services/homeApi.dart
@@ -459,7 +459,7 @@ class HomeApi {
 
   static Future<String> createOrder({
     String? date,
-    int? total_price,
+    double? total_price,
     String? shipping_type,
     String? payment_term,
     String? note,
